rename and correct FWave.h
